// src/app/static/src/generated.d.ts
export interface GeoJSONObject {
  type: "FeatureCollection";
  crs?: {
    [k: string]: any;
  };
  name?: string;
  features: {
    [k: string]: any;
  }[];
  [k: string]: any;
}
// ...
export interface ShapeResponse {
  hash: string;
  filename: string;
  fromADR?: boolean;
  resource_url?: string | null;
  type: "shape";
  data: GeoJSONObject;
  filters: {
    level_labels?: {
      id: number;
      area_level_label: string;
      display: boolean;
    }[];
    regions?: {
      label: string;
      id: string;
      children?: {
        [k: string]: any;
      }[];
    };
    [k: string]: any;
  };
}